+++ .github/workflows/test.yaml
@@ -39,7 +39,7 @@ jobs:
 
       - name: Copy web assets
         run: |
-          cp admin/dist/bundle.js ./internal/admin/handlers/web/assets/js/
+          cp admin/dist/bundle.js ./internal/web/handlers/assets/js/
 
       - name: Cache Go modules
         uses: actions/cache@v1
