+++ .gitignore
@@ -1,8 +1,8 @@
 # build
 bin/*
 dist/
-internal/admin/handlers/web/assets/fonts/*
-internal/admin/handlers/web/assets/js/*
+internal/web/handlers/assets/fonts/*
+internal/web/handlers/assets/js/*
 internal/**/*.pb.go
 
 # NVM
