+++ internal/web/handlers/list.go
