+++ internal/web/handlers/search.go
