+++ internal/web/setup_test.go
@@ -1,4 +1,4 @@
-package admin
+package web
 
 import (
 	"os"
